src/Analysis/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
awkward==2.6.9
awkward_cpp==39
bokeh==3.4.3
boost-histogram==1.3.2
clang==17.0.3
cmsstyle==0.4.0
coffea==2024.10.0
contourpy==1.3.0
correctionlib==2.6.4
cramjam==2.9.0
dask==2024.8.0
dask-awkward==2024.9.0
dask-histogram==2024.9.1
dask-jobqueue==0.8.5
distributed==2024.8.0
dxr==0.1
frozenlist==1.5.0
fsspec_xrootd==0.4.0
kiwisolver==1.4.7
llvmlite==0.43.0
multidict==6.1.0
numba==0.60.0
pacparser==1.0.0
propcache==0.2.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
sortedcontainers==2.4.0
tblib==3.0.0
toolz==1.0.0
tqdm==4.67.0
typing_extensions==4.12.2
uproot==5.5.0
vector==1.5.2
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.17.1
zict==3.0.0
